Rename DefaultNavigation to Navigation
